% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawSingleGeneStructure.R
\name{DrawSingleGeneStructure}
\alias{DrawSingleGeneStructure}
\title{to draw gene structure of each single gene.}
\usage{
DrawSingleGeneStructure(
  gffFile = NULL,
  haploResults = NULL,
  size = 5,
  lcolor = "red",
  pdfWidth = 16,
  pdfHeight = 4,
  ...
)
}
\arguments{
\item{gffFile:}{generic feature format file of certain genome, and this file defines the start and end position of specific gene, which should contain at least 9 columns such as "seqid", "source", "type", "start", "end", "score", "strand", "phase" and "attributes";}

\item{haploResults:}{haplotype analysis results, could be either a folder stores gene list files or a single file of gene list;}

\item{size:}{height of gene body, default value is 5;}

\item{lcolor:}{color of the middle line, default value is "red".}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this package, you could use the function of DrawSingleGeneStructure to draw gene structure of each single gene.
}
\details{
This is some description of this function.

see above
}
\examples{
DrawSingleGeneStructure(gffFile = "./Wm82.a2.v1.gene.gff", haploResults = "./Haplotype.Analysis/",size = 5, lcolor= "red")
}
