% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geno2Fasta.R
\name{Geno2Fasta}
\alias{Geno2Fasta}
\title{to transform geno data into fasta sequence}
\usage{
Geno2Fasta(genoSiteFile = NULL, nthreads = 1, ...)
}
\arguments{
\item{genositeFile:}{geno site classification folder;}

\item{nthreads:}{to define how many cores of this computer you would like to run the task. Make sure not to excede the maximum of this computer. Default value is 1.}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this package, you could use the function of Geno2Fasta to transform geno data into fasta sequence.
}
\details{
This is some description of this function.

see above
}
\examples{
Geno2Fasta(genositeFile = "./Geno.Site.Classification/", nthreads = 4)
}
