% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geno2Hmp.R
\name{Geno2Hmp}
\alias{Geno2Hmp}
\title{transforming the .vcf or the .geno into .hapmap format}
\usage{
Geno2Hmp(genoFile = NULL, nthreads = 1)
}
\arguments{
\item{genoFile:}{the files in vcf format or .geno format you want to transform;}

\item{nthreads:}{to define how many cores of this computer you would like to run the task. Make sure not to excede the maximum of this computer. Default value is 1.}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this package, you could transform the .vcf or the .geno into .hapmap format.
}
\details{
This is some description of this function.

see above
}
\examples{
Geno2Hmp(genoFile = "./vcfFile/",nthreads = 4)
}
