% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoExtractFromHapmap.R
\name{GenoExtractFromHapmap}
\alias{GenoExtractFromHapmap}
\title{to extract samples from raw.hapmap data}
\usage{
GenoExtractFromHapmap(trait = NULL, categoryFile = NULL, genoFile = NULL, ...)
}
\arguments{
\item{trait:}{the trait you are working with;}

\item{categoryFile:}{file stores category of each sample;}

\item{genoFile:}{a folder stores the hapmap file of each chromosome, or a single hapmap file of specific chromosome.}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this package, you could use functions of GenoExtractFromHapmap to extract samples from raw.hapmap data
}
\details{
This is some description of this function.

see above
}
\examples{
GenoExtractFromHapmap(trait = "Trait",categoryFile = "./Trait.category.csv",genoFile = "./soy_hapmap/")
}
