% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoSiteClassification.R
\name{GenoSiteClassification}
\alias{GenoSiteClassification}
\title{to classify SNP sites according to relative position to a certain gene}
\usage{
GenoSiteClassification(annoFile = NULL, hapmapFile = NULL, ...)
}
\arguments{
\item{annoFile:}{a SNP annotation folder from the fuction SNPAnnotation;}

\item{hapmapFile:}{regional hapmap file extracted from hapmap, could be either a file or a folder storing these type of files;}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this package, you could use functions of GenoSiteClassification to classify SNP sites according to relative position to a certain gene
}
\details{
This is some description of this function.

see above
}
\examples{
GenoSiteClassification(annoFile = "./SNP.Annotation/", hapmapFile = "./Hapmap.Regional.Extracted.Files/")
}
