% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploAssistant.R
\name{HaploAssistant}
\alias{HaploAssistant}
\title{haplotype analysis for specific genes.}
\usage{
HaploAssistant(
  categoryTransforming = F,
  GenoTransforming = F,
  GenoExtract = T,
  GeoPlotting = T,
  trait = NULL,
  SNPAnnoFile = NULL,
  categoryFile = NULL,
  gffFile = NULL,
  genomeFasta = NULL,
  phenoFile = NULL,
  geneModel = NULL,
  genoFile = NULL,
  mapData = NULL,
  haploMin = 10,
  pdfWidth = 10,
  pdfHeight = NULL,
  plotMargin = NULL,
  nthreads = 1,
  geneScope = 2000,
  size = 5,
  lcolor = "red",
  type = "countries",
  region = "China",
  mapResolution = 50,
  mapBorderSize = 0.1,
  longLim = NULL,
  latLim = NULL,
  landColor = "lightyellow",
  oceanColor = "white",
  pieMax = 2,
  expand = T,
  dataFiltration = NULL,
  pieLinetype = 2,
  pieLinesize = 0,
  pieBorderColor = NA,
  pieAlpha = 0.8,
  label_graticule = "EW",
  MappingDataType = "haplotype"
)
}
\arguments{
\item{categoryTransforming:}{determine whether to convert latitude and longitude data format;}

\item{GenoTransforming:}{determine whether to convert the file format from .vcf/.geno to .hapmap. Default is 'F'.}

\item{GenoExtract:}{determine whether to extract sample data from original Hapmap file. Default is "T".}

\item{GeoPlotting:}{determine whether to conduct geographic mapping. Default is "T".}

\item{trait:}{the trait you are working with;}

\item{SNPAnnoFile:}{SNP annotation files for users self sequencing;}

\item{categoryFile:}{file stores category of each sample;}

\item{gffFile:}{generic feature format file of certain genome. This file defines the start and end position of specific gene, it should contains at least "seqid", "source", "type", "start", "end", "score", "strand", and "attributes";}

\item{genomeFasta:}{genomic sequence in fasta format, which is required;}

\item{phenoFile:}{trait value of each sample;}

\item{geneModel:}{gene model you want to perform haplotype analysis, or a file with a column named "Gene_ID";}

\item{genoFile:}{a folder stores the hapmap file of each chromosome, or a single hapmap file of specific chromosome;}

\item{mapData:}{geographic data in 'shp' or 'json' format. If 'mapData' is NULL, map data would be downloaded from rnaturalearth (http://www.naturalearth.com).}

\item{haploMin:}{the minimum samples of a specific haplotype taking into account;}

\item{pdfWidth:}{numeric value to indicate the width of pdf file, default is 15;}

\item{pdfHeight:}{numeric value to indicate the height of pdf file, default is 10;}

\item{plotMargin:}{adjust graph edge margins;}

\item{nthreads:}{to define how many cores of this computer you would like to run the task. Make sure not to excede the maximum of this computer. Default value is 1;}

\item{geneScope:}{a region to define the upstream and downstream of a gene, genomic position out of this scope would be defined as intergenic. Default value is 2000;}

\item{size:}{height of gene body, default value is 5;}

\item{lcolor:}{color of the middle line, default value is "red";}

\item{type:}{could be either coastline or countries, default is countries;}

\item{region:}{the region to display on the map, could be "world", "China", or other countries, default is 'China';}

\item{mapResolution:}{value of 10, 50 and 110, or string of 'small', 'medium', or 'large', default is 50;}

\item{mapBorderSize:}{border size of geographic mapping, default is 0.1.}

\item{longLim:}{limitation of longitude, default is c(-180.00, 180.00);}

\item{latLim:}{limitation of latitude, default is c(-90.00, 90.00);}

\item{landColor:}{the color of land, default is lightyellow;}

\item{oceanColor:}{the color of ocean, default is white;}

\item{pieMax:}{numeric value to indicate the maximum size of pie plot on geographic map, default is 2;}

\item{expand:}{make the axes not overlap with each other, default is "T";}

\item{pieLinetype:}{line type inherited from ggplot2, default is 2;}

\item{pieLinesize:}{line size inherited from ggplot2, default is 0;}

\item{pieBorderColor:}{border color of pie plot, default is NA;}

\item{pieAlpha:}{numeric value to define transparency of pie plot, default is 0.8;}

\item{label_graticule:}{parameter of coord_sf(), which is a character vector indicating which graticule lines should be labeled where. It could be 'N', 'S', 'W', 'E', or combination of these characters. Default is 'EW'.}

\item{MappingDataType:}{data types of geographical mapping, default is "haplotype";}

\item{dataFiltration:}{logic value to indicatee whether to filtrate data by 'China', or other keywords, default is NULL.}
}
\value{
Files extracted and stored in the folder created by default. File types of hapmap, log, csv, fasta, haplotype, pdf.
}
\description{
By using this package, you could use functions of HaploAssistant to haplotype analysis for specific genes.
}
\details{
This is some descriptions of this function.

see above
}
\examples{
HaploAssistant(categoryTransforming = F, GenoTransforming = F, GenoExtract = T, GeoPlotting = T,
                         trait = "Trait", categoryFile = "./Trait.category.csv",
                         gffFile = "./Wm82.a2.v1.gene.gff", genomeFasta = "./Gmax_275_v2.0.fa",
                         phenoFile ="./Trait.csv", geneModel="Glyma.18G056600",
                         genoFile = "./Hapmap_Gm18.txt", mapData = "./China.json")
}
