% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaplotypePlotting.R
\name{HaplotypePlotting}
\alias{HaplotypePlotting}
\title{to compare phenotype difference between haplotypes}
\usage{
HaplotypePlotting(
  haploResults = NULL,
  haploMin = 10,
  pdfWidth = 10,
  pdfHeight = 10,
  plotMargin = 2,
  nthreads = 1,
  ...
)
}
\arguments{
\item{haploResults:}{haplotype analysis results, could be either a folder stores gene list files or a single file of gene list;}

\item{haploMin:}{the minimum samples of a specific haplotype taking into account;}

\item{plotMargin:}{adjust graph edge margins, default is 2;}

\item{nthreads:}{to define how many cores of this computer you would like to run the task. Make sure not to excede the maximum of this computer. Default value is 1.}
}
\value{
files of pdf
}
\description{
By using this package, you could use function of HaplotypePlotting to compare phenotype difference between haplotypes
}
\details{
This is some description of this function.

see above
}
\examples{
HaplotypePlotting(haploResults = "./Haplotype.Analysis/", nthreads = 6)
}
