% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongLatTransforming.R
\name{LongLatTransforming}
\alias{LongLatTransforming}
\title{to transform longitude or latitude into different format, either from dms to dd, or from dd to dms}
\usage{
LongLatTransforming(longLat = NULL, ...)
}
\arguments{
\item{longLat:}{input data to be transformed. It could be either value with or without 'E|N|W|S' to indicate direction, or file including longitude or latitude column. For example,}
}
\value{
value or file, depends on the input data.
}
\description{
By using this package, you could use functions of LongLatTransforming to transform longitude or latitude into different format, either from dms to dd, or from dd to dms
}
\details{
This is some description of this function.

In the LongLatTransforming, the input data could either be a regular format, or files containing longitude and latitude.
}
\examples{
LongLatTransforming("-116°25'W")

}
