% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NexForPopART.R
\name{NexForPopART}
\alias{NexForPopART}
\title{generate .nex file stores DATA and TRAITS for haplotype network inference using PopART}
\usage{
NexForPopART(haploResults = NULL, categoryFile = "none", ...)
}
\arguments{
\item{haploResults:}{haplotype category and frequency of haplotype analysis results, could be either a folder stores gene list files or a single file of gene list;}

\item{categoryFile:}{longitude and latitude of each location.}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this function, you could generate .nex file stores DATA and TRAITS for haplotype network inference using PopART
}
\details{
This is some description of this function.

see above
}
\examples{
NexForPopART(haploResults = "./Haplotype.Analysis/", categoryFile = "./Trait.category.csv")
}
