% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionalSNPextractionFromHapmap.R
\name{RegionalSNPextractionFromHapmap}
\alias{RegionalSNPextractionFromHapmap}
\title{extract SNP genotype of of a certain regionFiles either to specific trait or not from hapmap files}
\usage{
RegionalSNPextractionFromHapmap(
  sampleHapmapFile = NULL,
  gffFile = NULL,
  geneModel = NULL,
  ...
)
}
\arguments{
\item{sampleHapmapFile:}{hapmap file after extracting samples, could be either a single hapmap file or a folder stores hapmap files;}

\item{gffFile:}{generic feature format file of certain genome, and this file defines the start and end position of specific gene, which should contain at least 9 columns such as "seqid", "source", "type", "start", "end", "score", "strand", "phase" and "attributes".}

\item{geneModel:}{gene model you want to perform haplotype analysis, or a file that stores gene names with the column names "Gene_ID".}
}
\value{
files extracted and stored in the folder created by default
}
\description{
By using this function, you could extract SNP genotype of of a certain regionFiles either to specific trait or not from hapmap files
}
\details{
This is some description of this function.

see above
}
\examples{
RegionalSNPextractionFromHapmap(sampleHapmapFile = "./Trait_geno.extracted.from.hapmap/",
                                          gffFile = "./Wm82.a2.v1.gene.gff",
                                          geneModel = "Glyma.18G056600")
}
