% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPAnnotation.R
\name{SNPAnnotation}
\alias{SNPAnnotation}
\title{to annotate where a specific SNP is relative to the nearest gene or within certain region of a gene.}
\usage{
SNPAnnotation(
  genomeFasta = NULL,
  gffFile = NULL,
  hapmapFile = NULL,
  geneScope = 2000,
  nthreads = 1,
  ...
)
}
\arguments{
\item{genomeFasta:}{genomic sequence in fasta format, which is required;}

\item{gffFile:}{generic feature format file of certain genome, and this file defines the start and end position of specific gene, which should contain at least 9 columns such as "seqid", "source", "type", "start", "end", "score", "strand", "phase" and "attributes".}

\item{hapmapFile:}{sample segment hapmap file, could be either a file or a folder storing these type of files;}

\item{geneScope:}{a region to define the upstream and downstream of a gene, genomic position out of this scope would be defined as intergenic. Default value is 2000;}

\item{nthreads:}{to define how many cores of this computer would be used to process the task. Default is 1. Make sure not to excede the maximum cores of this computer;}
}
\value{
files extracted and stored in the folder created by default.
}
\description{
By using this package, you could use functions of SNPAnnotation to annotate where a specific SNP is relative to the nearest gene or within certain region of a gene.
}
\details{
This is some description of this function.

see above
}
\examples{
SNPAnnotation(genomeFasta = "./Gmax_275_v2.0.fa", gffFile = "./Wm82.a2.v1.gene.gff", hapmapFile = "./Hapmap.regional.extracted.files/")
}
